Refactor out archetypes-code into module
